JWT implementation & small JSDoc adjust
The JSDoc needs to follow one of the specific plugins, now we want to be able to refer to modules/classes outside of our current module/class.
To do this we're gonna need some kind of plugin to support it, as far as I can tell jsdoc3 doesn't support it. So syntax is subject to change dependent
on how the plugin handles {@link name Class} or whatever syntax it uses.

Added basic JWT to the cookie in response to a confirmed login. Expiry is 10 seconds for testing, left a comment with the intended value it should be
at once development on the CMS auth is complete. (2 days)