Card endpoint initial done
Now we just need the query string to set an out/offset for the getting of the cards. I think just filtering through the entire list will be fine, we don't need mongo to do this for us.
There's not very much data in a card, maximum 105 chars in the text too...